+++ app/src/components/layouts/LandingLayout/LandingLayout.tsx
@@ -1,15 +1,18 @@
 import { Center } from '@/components/common';
 import styled from '@emotion/styled';
 import { Helmet } from 'react-helmet-async';
+import { Outlet } from 'react-router-dom';
 
-export const LandingLayout = ({ children }: React.PropsWithChildren) => {
+export const LandingLayout = () => {
   return (
     <>
       <Helmet>
         <meta name="theme-color" content="#212222" />
       </Helmet>
       <LandingPageLayout>
-        <Center h="100vh">{children}</Center>
+        <Center h="100vh">
+          <Outlet />
+        </Center>
       </LandingPageLayout>
     </>
   );
